backend/requirements.txt
langchain-openai
langchain_ollama
fastapi
uvicorn
langchain
pydantic
transformers
chromadb
sentence-transformers
arxiv
python-dotenv
langchain-community
langchain-chroma
python-multipart
